requirements.txt
inspect-ai>=0.3.201
inspect-evals[swe_bench]>=0.6.0
anthropic>=0.86.0
datasets>=4.8.4
python-dotenv>=1.2.2
